+++ admin/sub_categories.php
@@ -12,14 +12,16 @@ if (isset($_GET['type']) && $_GET['type'] != '') {
 		} else {
 			$status = '0';
 		}
-		$update_status_sql = "update categories set status='$status' where id='$id'";
+		$update_status_sql = "update sub_categories set status='$status' where id='$id'";
 		mysqli_query($con, $update_status_sql);
+		set_flash('msg', 'Sub category status updated successfully ✅');
 	}
 
 	if ($type == 'delete') {
 		$id = get_safe_value($con, $_GET['id']);
-		$delete_sql = "delete from categories where id='$id'";
+		$delete_sql = "delete from sub_categories where id='$id'";
 		mysqli_query($con, $delete_sql);
+		set_flash('msg', 'Sub category deleted successfully ✅');
 	}
 }
 
@@ -67,7 +69,7 @@ $res = mysqli_query($con, $sql);
 												}
 												echo "<span class='badge badge-edit'><a href='manage_sub_categories.php?id=" . $row['id'] . "'>Edit</a></span>&nbsp;";
 
-												echo "<span class='badge badge-delete'><a href='?type=delete&id=" . $row['id'] . "'>Delete</a></span>";
+												echo "<span class='badge badge-delete'><a href='?type=delete&id=" . $row['id'] . "'onclick=\"return confirm('Are you sure you want to delete this sub category?')\">Delete</a></span>";
 
 												?>
 											</td>
@@ -76,6 +78,11 @@ $res = mysqli_query($con, $sql);
 								</tbody>
 							</table>
 						</div>
+						<?php if ($msg = get_flash('msg')) { ?>
+							<div id="myToast" class="myToast__msg">
+								<?php echo $msg; ?>
+							</div>
+						<?php } ?>
 					</div>
 				</div>
 			</div>
